+++ ablateCore/CMakeLists.txt
@@ -1,10 +1,10 @@
 # Include PETSc and MPI
-target_include_directories(ablateLibrary PUBLIC ${PETSc_INCLUDE_DIRS})
-target_link_libraries(ablateLibrary PUBLIC ${PETSc_LIBRARIES})
-target_link_directories (ablateLibrary PUBLIC ${PETSc_LIBRARY_DIRS})
+target_include_directories(ablateCore PUBLIC ${PETSc_INCLUDE_DIRS})
+target_link_libraries(ablateCore PUBLIC ${PETSc_LIBRARIES})
+target_link_directories (ablateCore PUBLIC ${PETSc_LIBRARY_DIRS})
 
 # Include code
-target_sources(ablateLibrary
+target_sources(ablateCore
         PUBLIC
         mesh.c
         mesh.h
@@ -13,7 +13,7 @@ target_sources(ablateLibrary
         )
 
 # Allow public access to the header files in the directory
-target_include_directories(ablateLibrary PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(ablateCore PUBLIC ${CMAKE_CURRENT_LIST_DIR})
 
 # Include the code in any subdirectory
 add_subdirectory(flow)
