+++ ablateCore/flow/CMakeLists.txt
@@ -1,6 +1,6 @@
 
 # Include code
-target_sources(ablateLibrary
+target_sources(ablateCore
         PUBLIC
         flow.c
         flow.h
@@ -11,4 +11,4 @@ target_sources(ablateLibrary
         )
 
 # Allow public access to the header files in the directory
-target_include_directories(ablateLibrary PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(ablateCore PUBLIC ${CMAKE_CURRENT_LIST_DIR})
