+++ ablateCore/particles/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Include code
-target_sources(ablateLibrary
+target_sources(ablateCore
         PUBLIC
         particles.c
         particles.h
@@ -11,4 +11,4 @@ target_sources(ablateLibrary
 add_subdirectory(initializers)
 
 # Allow public access to the header files in the directory
-target_include_directories(ablateLibrary PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(ablateCore PUBLIC ${CMAKE_CURRENT_LIST_DIR})
