+++ ablateCore/particles/initializers/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Include code
-target_sources(ablateLibrary
+target_sources(ablateCore
         PUBLIC
         particleInitializer.c
         particleInitializer.h
@@ -10,4 +10,4 @@ target_sources(ablateLibrary
         )
 
 # Allow public access to the header files in the directory
-target_include_directories(ablateLibrary PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(ablateCore PUBLIC ${CMAKE_CURRENT_LIST_DIR})
