+++ tests/ablateCore/outputs/incompressible_2d_tri_p2_p1_p1_tconv
