--- tests/testFixtures/CMakeLists.txt
@@ -1,7 +1,0 @@
-target_sources(tests
-        PRIVATE
-        MpiTestFixture.hpp
-        MpiTestFixture.cpp
-        MpiTestParamFixture.hpp
-        MpiTestEventListener.h
-        )
